requirements.txt
streamlit
google-generativeai
pandas
numpy
pyyaml
requests
fastapi
uvicorn[standard]
pydantic
python-dotenv
firebase-admin
beautifulsoup4